tentativas não deram certas
