+++ whatslab/src/App.js
@@ -21,8 +21,10 @@ function App() {
   return (
     <Container>
       <HeaderWhatsLab />
-      <MainWhatsLab></MainWhatsLab>
+      <MainWhatsLab />
       <ListMessenger />
+
+
     </Container>
 
   );
